Centralise Theme Methods
