+++ src/Traits/ComponentUtilities.php
@@ -16,8 +16,6 @@ trait ComponentUtilities
 
     public array $table = [];
 
-    public ?string $theme = null;
-
     protected Builder $builder;
 
     protected $model;
@@ -60,8 +58,8 @@ trait ComponentUtilities
     public function mountComponentUtilities(): void
     {
         // Sets the Theme - tailwind/bootstrap
-        if (is_null($this->theme)) {
-            $this->setTheme();
+        if (!isset($this->theme) || is_null($this->theme)) {
+            $this->setTheme(config('livewire-tables.theme', 'tailwind'));
         }
         $this->generateDataTableFingerprint();
