+++ src/Traits/HasAllTraits.php
@@ -2,11 +2,14 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits;
 
+use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasTheme;
+
 trait HasAllTraits
 {
     // Note Specific Order Below!
     use WithTableHooks;
     use WithLoadingPlaceholder;
+    use HasTheme;
     use ComponentUtilities,
         WithActions,
         WithData,
