+++ src/Views/Traits/Configuration/ColumnConfiguration.php
@@ -84,13 +84,6 @@ trait ColumnConfiguration
         $this->displayColumnLabel = $status;
     }
 
-    public function setTheme(string $theme): self
-    {
-        $this->theme = $theme;
-
-        return $this;
-    }
-
     public function setHasTableRowUrl(bool $hasTableRowUrl): self
     {
         $this->hasTableRowUrl = $hasTableRowUrl;
